Domenica 20

+++ src/main/java/ai/polito/lab2/demo/controllers/ReservationController.java
@@ -2,7 +2,6 @@ package ai.polito.lab2.demo.controllers;
 
 //import ai.polito.lab2.demo.AppConfig;
 
-import ai.polito.lab2.demo.Entity.Child;
 import ai.polito.lab2.demo.Entity.Reservation;
 import ai.polito.lab2.demo.Entity.Stop;
 import ai.polito.lab2.demo.Service.*;
@@ -41,9 +40,6 @@ public class ReservationController {
 
     @Autowired
     private StopService stopService;
-
-    //@Autowired
-    //private EmailSenderService emailSenderService;
 
     @Autowired
     private ReservationService reservationService;
@@ -102,7 +98,7 @@ public class ReservationController {
      * @throws JsonProcessingException
      * @throws ParseException
      */
-    @Secured("ROLE_MULE")
+    @Secured({"ROLE_MULE","ROLE_ADMIN"})
     @RequestMapping(value = "/reservations/add/{id_linea}/{data}", method = RequestMethod.POST, consumes = MediaType.APPLICATION_JSON_VALUE, produces = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation("RICHIESTA per aggiungere un bambino non prenotato ma presente alla fermata")
     public ResponseEntity createNotBooked(@PathVariable int id_linea, @PathVariable long data, @RequestBody ReservationVM reservationVM) throws JsonProcessingException, ParseException {
@@ -133,7 +129,7 @@ public class ReservationController {
      * @throws JsonProcessingException
      * @throws ParseException
      */
-    @Secured({"ROLE_MULE"})
+    @Secured({"ROLE_MULE","ROLE_ADMIN"})
     @RequestMapping(value = "/reservations/{id_fermata}/{data}", method = RequestMethod.PUT)
     @ApiOperation("RICHIESTA per confermare o meno presenza del bambino")
     public ResponseEntity confirmPresence(@PathVariable final String id_fermata, @PathVariable long data, @RequestBody childConfirmVM childConfirmVM) throws JsonProcessingException, ParseException {
@@ -193,7 +189,7 @@ public class ReservationController {
      * @param reservation_id id reservation da aggiornare
      * @return
      */
-    @Secured("ROLE_USER")
+   /* @Secured("ROLE_USER")
     @RequestMapping(value = "/reservations/{nome_linea}/{data}/{reservation_id}", method = RequestMethod.PUT, consumes = MediaType.APPLICATION_JSON_VALUE)
     @ApiOperation("Per aggiornare una reservation")
     public ResponseEntity update(@RequestBody ReservationVM reservationVM, @PathVariable String nome_linea, @PathVariable long data, @PathVariable final ObjectId reservation_id) {
@@ -213,7 +209,6 @@ public class ReservationController {
             updatedReservation.setDate(data);
         }
 
-        // TODO per il lab 5 non serve ma controllare qui la corrispondenza tra ChildReservationVM e il Child
         if (!reservationVM.getChildID().toString().isEmpty()) {
             updatedReservation.setChildID(childID);
         }
@@ -233,7 +228,7 @@ public class ReservationController {
 
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
-
+*/
     /**
      * Va a controllare se è possibile cambiare o eliminare la prenotazione
      * @param nowTimeStamp ora odierna
@@ -298,7 +293,7 @@ public class ReservationController {
      * @param reservation_id id della reservation da eliminare
      * @return
      */
-    @Secured({"ROLE_USER", "ROLE_MULE"})
+    @Secured({"ROLE_USER"})
     @RequestMapping(value = "/reservations/{reservation_id}", method = RequestMethod.DELETE)
     @ApiOperation("Eliminazione della reservation")
     public ResponseEntity delete(@ApiParam("Id della prenotazione da eliminare")@PathVariable ObjectId reservation_id) {
@@ -313,7 +308,7 @@ public class ReservationController {
         reservationService.delete(reservation_id);
         return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
-
+/*
     @Secured({"ROLE_USER", "ROLE_MULE"})
     @RequestMapping(value = "/reservations/{reservation_id}", method = RequestMethod.GET)
     public ResponseEntity<Reservation> getPeople(@PathVariable ObjectId reservation_id) throws JsonProcessingException {
@@ -321,9 +316,14 @@ public class ReservationController {
 
         return new ResponseEntity<>(request,HttpStatus.OK);
     }
-
-   // TODO: prendersi anche l'ora della fermata
-    @Secured({"ROLE_USER", "ROLE_MULE"})
+*/
+    /**
+     * Funzione per il calendario del genitore
+     * @param family_name cognome della famiglia
+     * @return tutte le prenotazioni per quella famiglia
+     * @throws JsonProcessingException
+     */
+    @Secured({"ROLE_USER"})
     @RequestMapping(value = "/reservations", method = RequestMethod.GET)
     public ResponseEntity getChildReservation(@RequestParam (required = true) String family_name) throws JsonProcessingException {
         System.out.println("family_name :" + family_name);
@@ -332,6 +332,12 @@ public class ReservationController {
         return ok().body(reservationCalendarVMS);
     }
 
+    /**
+     * Scaricare tutte le prenotazioni per un bimbo per l'admin
+     * @param childID id del bambino per cui si vogliono prendere tutte le prenotazioni
+     * @return
+     * @throws JsonProcessingException
+     */
     @Secured({"ROLE_SYSTEM_ADMIN"})
     @RequestMapping(value = "/reservations/child/{childID}", method = RequestMethod.GET)
     public ResponseEntity getChildListReservations(@PathVariable ObjectId childID) throws JsonProcessingException {
@@ -343,13 +349,28 @@ public class ReservationController {
 
 
     // A livello stilistico era meglio farlo con il PathParam ma funziona anche così
+
+    /**
+     *
+     * @param id
+     * @return
+     * @throws JsonProcessingException
+     */
     @Secured({"ROLE_USER", "ROLE_MULE"})
     @RequestMapping(value = "/reservations", method = RequestMethod.DELETE)
     public ResponseEntity deleteChildReservation(@RequestParam (required = true) ObjectId id) throws JsonProcessingException {
 
+        Reservation updatedReservation = reservationService.findReservationById(id);
+        long nowTimeStamp = getCurrentTimeStamp();
+        Stop stop = stopService.findStopbyId(updatedReservation.getStopID());
+
+        if (checkTimestamp(nowTimeStamp,updatedReservation.getDate(),stop))
+        {
+            return new ResponseEntity<>("Errore non è possibile cancellare la prenotazione",HttpStatus.BAD_REQUEST);
+        }
         reservationService.delete(id);
 
-        return noContent().build();
+        return new ResponseEntity(HttpStatus.NO_CONTENT);
     }
